+++ src/main/java/com/arangodb/entity/arangosearch/ArangoSearchProperties.java
@@ -27,6 +27,8 @@ import java.util.Collection;
 /**
  * @author Mark Vollmary
  * @author Heiko Kernbach
+ * @author Michele Rastelli
+ * @see <a href="https://www.arangodb.com/docs/stable/http/views-arangosearch.html">API Documentation</a>
  */
 public class ArangoSearchProperties {
 
@@ -36,11 +38,14 @@ public class ArangoSearchProperties {
     private ConsolidationPolicy consolidationPolicy;
     private final Collection<PrimarySort> primarySorts;
     private final Collection<CollectionLink> links;
+    private ArangoSearchCompression primarySortCompression;
+    private final Collection<StoredValue> storedValues;
 
     public ArangoSearchProperties() {
         super();
         links = new ArrayList<>();
         primarySorts = new ArrayList<>();
+        storedValues = new ArrayList<>();
     }
 
     public Long getCommitIntervalMsec() {
@@ -90,4 +95,32 @@ public class ArangoSearchProperties {
     public void addPrimarySort(final PrimarySort... primarySorts) {
         this.primarySorts.addAll(Arrays.asList(primarySorts));
     }
+
+    /**
+     * @return Defines how to compress the primary sort data (introduced in v3.7.0). ArangoDB v3.5 and v3.6 always
+     * compress the index using LZ4.
+     * @since ArangoDB 3.7
+     */
+    public ArangoSearchCompression getPrimarySortCompression() {
+        return primarySortCompression;
+    }
+
+    public void setPrimarySortCompression(ArangoSearchCompression primarySortCompression) {
+        this.primarySortCompression = primarySortCompression;
+    }
+
+    /**
+     * @return An array of objects to describe which document attributes to store in the View index. It can then cover
+     * search queries, which means the data can be taken from the index directly and accessing the storage engine can be
+     * avoided.
+     * @since ArangoDB 3.7
+     */
+    public Collection<StoredValue> getStoredValues() {
+        return storedValues;
+    }
+
+    public void addStoredValues(final StoredValue... storedValues) {
+        this.storedValues.addAll(Arrays.asList(storedValues));
+    }
+
 }