+++ ChangeLog.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+## [6.7.0_PREVIEW_3.7.1-alpha.1] - 2020-05-22
+
+- added support of `DisjointSmartGraphs` and `SatelliteGraphs` (ArangoDB v3.7)
+- added support of `storedValues` in `ArangoSearchProperties` (ArangoDB v3.7)
+- added support of `primarySortCompression` in `ArangoSearchProperties` (ArangoDB v3.7)
+- added support of `overwriteMode` on document creation, to allow `insert-ignore`, `insert-replace` and `insert-update` (ArangoDB v3.7)
+- added support of `mergeObjects` for insert document with `overwriteMode: update` (ArangoDB v3.7)
+- velocypack v2.3.1
+
 ## [6.6.3] - 2020-05-06
 
 - velocypack v2.3.1
@@ -14,6 +23,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - bugfix VelocyJack deserialization
 - bugfix `allowImplicit` parameter in stream transactions
+
+## [6.7.0_PREVIEW_3.7.0-alpha.2_0] - 2020-03-24
+
+- added `overwriteMode` parameter to support insert-update (ArangoDB v3.7)
+- satellite graphs support (ArangoDB v3.7)
+- schema validation (ArangoDB v3.7)
 - added `peakMemoryUsage` to aql statistics
 
 ## [6.6.1] - 2020-03-18